+++ viadot/sources/__init__.py
@@ -5,6 +5,11 @@ from .supermetrics import Supermetrics
 from .cloud_for_customers import CloudForCustomers
 from .sharepoint import Sharepoint
 
+try:
+    from .sap_rfc import SAPRFC
+except ImportError:
+    pass
+
 # APIS
 from .uk_carbon_intensity import UKCarbonIntensity
 from .sqlite import SQLite