+++ viadot/tasks/__init__.py
@@ -24,3 +24,8 @@ from .sqlite import SQLiteInsert, SQLiteSQLtoDF, SQLiteQuery
 from .supermetrics import SupermetricsToCSV, SupermetricsToDF
 from .sharepoint import SharepointToDF
 from .cloud_for_customers import C4CReportToDF, C4CToDF
+
+try:
+    from .sap_rfc import SAPRFCToDF
+except ImportError:
+    pass